Move Dynamic field component to shared && change error message.
